Development

- Register all models on admin site
- Added cloudbuild config file and docker config file for CI/CD pipeline

+++ core/models.py
@@ -3,7 +3,7 @@ from pos.models import SalesItem
 from hrm.models import Employee
 
 class Commission(models.Model):
-    com_id = models.IntegerField(primary_key=True)
+    com_id = models.BigAutoField(primary_key=True)
     sales_item = models.ForeignKey(SalesItem, on_delete=models.PROTECT)
     emp = models.ForeignKey(Employee, on_delete=models.PROTECT)
     com_amt = models.DecimalField(max_digits=1000, decimal_places=10)
